+++ http-client/src/main/java/io/micronaut/http/client/netty/ConnectionManager.java
@@ -1,5 +1,6 @@
 package io.micronaut.http.client.netty;
 
+import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.annotation.Nullable;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpVersion;
@@ -11,6 +12,7 @@ import io.micronaut.http.netty.channel.ChannelPipelineListener;
 import io.micronaut.http.netty.stream.DefaultHttp2Content;
 import io.micronaut.http.netty.stream.Http2Content;
 import io.micronaut.http.netty.stream.HttpStreamsClientHandler;
+import io.micronaut.http.netty.stream.StreamingInboundHttp2ToHttpAdapter;
 import io.micronaut.scheduling.instrument.InvocationInstrumenter;
 import io.micronaut.websocket.exceptions.WebSocketSessionException;
 import io.netty.bootstrap.Bootstrap;
@@ -38,6 +40,7 @@ import io.netty.handler.codec.LineBasedFrameDecoder;
 import io.netty.handler.codec.http.DefaultHttpContent;
 import io.netty.handler.codec.http.FullHttpMessage;
 import io.netty.handler.codec.http.HttpClientCodec;
+import io.netty.handler.codec.http.HttpClientUpgradeHandler;
 import io.netty.handler.codec.http.HttpContent;
 import io.netty.handler.codec.http.HttpContentDecompressor;
 import io.netty.handler.codec.http.HttpMessage;
@@ -46,12 +49,18 @@ import io.netty.handler.codec.http.HttpUtil;
 import io.netty.handler.codec.http.LastHttpContent;
 import io.netty.handler.codec.http.websocketx.extensions.compression.WebSocketClientCompressionHandler;
 import io.netty.handler.codec.http2.DefaultHttp2Connection;
+import io.netty.handler.codec.http2.DelegatingDecompressorFrameListener;
+import io.netty.handler.codec.http2.Http2ClientUpgradeCodec;
 import io.netty.handler.codec.http2.Http2Connection;
+import io.netty.handler.codec.http2.Http2FrameListener;
 import io.netty.handler.codec.http2.Http2FrameLogger;
 import io.netty.handler.codec.http2.Http2Stream;
 import io.netty.handler.codec.http2.HttpToHttp2ConnectionHandler;
 import io.netty.handler.codec.http2.HttpToHttp2ConnectionHandlerBuilder;
+import io.netty.handler.codec.http2.InboundHttp2ToHttpAdapterBuilder;
 import io.netty.handler.logging.LoggingHandler;
+import io.netty.handler.ssl.ApplicationProtocolNames;
+import io.netty.handler.ssl.ApplicationProtocolNegotiationHandler;
 import io.netty.handler.ssl.SslContext;
 import io.netty.handler.ssl.SslHandler;
 import io.netty.handler.timeout.IdleStateEvent;
@@ -501,6 +510,152 @@ final class ConnectionManager {
     }
 
     /**
+     * Configures HTTP/2 for the channel when SSL is enabled.
+     *
+     * @param httpClientInitializer The client initializer
+     * @param ch                    The channel
+     * @param sslCtx                The SSL context
+     * @param host                  The host
+     * @param port                  The port
+     * @param connectionHandler     The connection handler
+     */
+    private void configureHttp2Ssl(
+        HttpClientInitializer httpClientInitializer,
+        @NonNull SocketChannel ch,
+        @NonNull SslContext sslCtx,
+        String host,
+        int port,
+        HttpToHttp2ConnectionHandler connectionHandler) {
+        ChannelPipeline pipeline = ch.pipeline();
+        // Specify Host in SSLContext New Handler to add TLS SNI Extension
+        pipeline.addLast(ChannelPipelineCustomizer.HANDLER_SSL, sslCtx.newHandler(ch.alloc(), host, port));
+        // We must wait for the handshake to finish and the protocol to be negotiated before configuring
+        // the HTTP/2 components of the pipeline.
+        pipeline.addLast(
+                ChannelPipelineCustomizer.HANDLER_HTTP2_PROTOCOL_NEGOTIATOR,
+                new ApplicationProtocolNegotiationHandler(ApplicationProtocolNames.HTTP_2) {
+
+            @Override
+            public void handlerRemoved(ChannelHandlerContext ctx) {
+                // the logic to send the request should only be executed once the HTTP/2
+                // Connection Preface request has been sent. Once the Preface has been sent and
+                // removed then this handler is removed so we invoke the remaining logic once
+                // this handler removed
+                final Consumer<ChannelHandlerContext> contextConsumer =
+                        httpClientInitializer.contextConsumer;
+                if (contextConsumer != null) {
+                    contextConsumer.accept(ctx);
+                }
+            }
+
+            @Override
+            protected void configurePipeline(ChannelHandlerContext ctx, String protocol) {
+                if (ApplicationProtocolNames.HTTP_2.equals(protocol)) {
+                    ChannelPipeline p = ctx.pipeline();
+                    if (httpClientInitializer.stream) {
+                        // stream consumer manages backpressure and reads
+                        ctx.channel().config().setAutoRead(false);
+                    }
+                    p.addLast(
+                            ChannelPipelineCustomizer.HANDLER_HTTP2_SETTINGS,
+                        httpClient.new Http2SettingsHandler(ch.newPromise())
+                    );
+                    httpClientInitializer.addEventStreamHandlerIfNecessary(p);
+                    httpClientInitializer.addFinalHandler(p);
+                    for (ChannelPipelineListener pipelineListener : pipelineListeners) {
+                        pipelineListener.onConnect(p);
+                    }
+                } else if (ApplicationProtocolNames.HTTP_1_1.equals(protocol)) {
+                    ChannelPipeline p = ctx.pipeline();
+                    httpClientInitializer.addHttp1Handlers(p);
+                } else {
+                    ctx.close();
+                    throw httpClient.customizeException(new HttpClientException("Unknown Protocol: " + protocol));
+                }
+                httpClientInitializer.onStreamPipelineBuilt();
+            }
+        });
+
+        pipeline.addLast(ChannelPipelineCustomizer.HANDLER_HTTP2_CONNECTION, connectionHandler);
+    }
+
+    /**
+     * Configures HTTP/2 handling for plaintext (non-SSL) connections.
+     *
+     * @param httpClientInitializer The client initializer
+     * @param ch                    The channel
+     * @param connectionHandler     The connection handler
+     */
+    private void configureHttp2ClearText(
+        HttpClientInitializer httpClientInitializer,
+        @NonNull SocketChannel ch,
+        @NonNull HttpToHttp2ConnectionHandler connectionHandler) {
+        HttpClientCodec sourceCodec = new HttpClientCodec();
+        Http2ClientUpgradeCodec upgradeCodec = new Http2ClientUpgradeCodec(ChannelPipelineCustomizer.HANDLER_HTTP2_CONNECTION, connectionHandler);
+        HttpClientUpgradeHandler upgradeHandler = new HttpClientUpgradeHandler(sourceCodec, upgradeCodec, 65536);
+
+        final ChannelPipeline pipeline = ch.pipeline();
+        pipeline.addLast(ChannelPipelineCustomizer.HANDLER_HTTP_CLIENT_CODEC, sourceCodec);
+        httpClientInitializer.settingsHandler = httpClient.new Http2SettingsHandler(ch.newPromise());
+        pipeline.addLast(upgradeHandler);
+        pipeline.addLast(new ChannelInboundHandlerAdapter() {
+            @Override
+            public void userEventTriggered(ChannelHandlerContext ctx, Object evt) throws Exception {
+                ctx.fireUserEventTriggered(evt);
+                if (evt instanceof HttpClientUpgradeHandler.UpgradeEvent) {
+                    httpClientInitializer.onStreamPipelineBuilt();
+                    ctx.pipeline().remove(this);
+                }
+            }
+        });
+        pipeline.addLast(ChannelPipelineCustomizer.HANDLER_HTTP2_UPGRADE_REQUEST, httpClient.new UpgradeRequestHandler(httpClientInitializer) {
+            @Override
+            public void handlerRemoved(ChannelHandlerContext ctx) {
+                final Consumer<ChannelHandlerContext> contextConsumer = httpClientInitializer.contextConsumer;
+                if (contextConsumer != null) {
+                    contextConsumer.accept(ctx);
+                }
+            }
+        });
+    }
+
+    /**
+     * Creates a new {@link HttpToHttp2ConnectionHandlerBuilder} for the given HTTP/2 connection object and config.
+     *
+     * @param connection    The connection
+     * @param configuration The configuration
+     * @param stream        Whether this is a stream request
+     * @return The {@link HttpToHttp2ConnectionHandlerBuilder}
+     */
+    @NonNull
+    private static HttpToHttp2ConnectionHandlerBuilder newHttp2ConnectionHandlerBuilder(
+            @NonNull Http2Connection connection, @NonNull HttpClientConfiguration configuration, boolean stream) {
+        final HttpToHttp2ConnectionHandlerBuilder builder = new HttpToHttp2ConnectionHandlerBuilder();
+        builder.validateHeaders(true);
+        final Http2FrameListener http2ToHttpAdapter;
+
+        if (!stream) {
+            http2ToHttpAdapter = new InboundHttp2ToHttpAdapterBuilder(connection)
+                    .maxContentLength(configuration.getMaxContentLength())
+                    .validateHttpHeaders(true)
+                    .propagateSettings(true)
+                    .build();
+
+        } else {
+            http2ToHttpAdapter = new StreamingInboundHttp2ToHttpAdapter(
+                    connection,
+                    configuration.getMaxContentLength()
+            );
+        }
+        return builder
+                .connection(connection)
+                .frameListener(new DelegatingDecompressorFrameListener(
+                        connection,
+                        http2ToHttpAdapter));
+
+    }
+
+    /**
      * Initializes the HTTP client channel.
      */
     protected class HttpClientInitializer extends ChannelInitializer<SocketChannel> {
@@ -558,7 +713,7 @@ final class ConnectionManager {
             if (httpVersion == HttpVersion.HTTP_2_0) {
                 final Http2Connection connection = new DefaultHttp2Connection(false);
                 final HttpToHttp2ConnectionHandlerBuilder builder =
-                        httpClient.newHttp2ConnectionHandlerBuilder(connection, configuration, stream);
+                        newHttp2ConnectionHandlerBuilder(connection, configuration, stream);
 
                 configuration.getLogLevel().ifPresent(logLevel -> {
                     try {
@@ -573,9 +728,9 @@ final class ConnectionManager {
                 HttpToHttp2ConnectionHandler connectionHandler = builder
                         .build();
                 if (sslContext != null) {
-                    httpClient.configureHttp2Ssl(this, ch, sslContext, host, port, connectionHandler);
+                    configureHttp2Ssl(this, ch, sslContext, host, port, connectionHandler);
                 } else {
-                    httpClient.configureHttp2ClearText(this, ch, connectionHandler);
+                    configureHttp2ClearText(this, ch, connectionHandler);
                 }
                 channelCustomizer.onInitialPipelineBuilt();
             } else {
